add `#[track_caller]` to methods
Before this change, any panics caused by the bounds checks point to the
location of the assert! macro inside this crate. This can be hard to
debug because there may be dozen or hundreds of uses of these methods
in a code base and the panic message doesn't tell which one caused the
panic. By attaching the `#[track_caller]` attribute, the panic message
will now point to the code that called the method.