add #[inline] to to_regular_range
Generic functions can already be inlined, but the compiler may not
necessarily do so. Explicitly adding the #[inline] attribute hints to
the compiler that it should inline the function. Given that the
function is a no-op in the best case and even in the worst case only
does two simple integer additions, this function should always be
inlined.